feat(gui): move the unsaved count and its popover into the save pill
The header's top-right count chip stated the same fact the pill
already narrates, in a second corner of the window. The pill's
count line is now the one opener: while the draft has attributed
rows it is a button whose popover renders the shared diff-rows
list (same renderer, same jump-to-control anchors); at zero rows
there is no list to open, so it stays plain text.

Retires home.unsaved.* across all catalogs (drop-key); adds
footer.unsaved.help for the next translation round. Surfacing
needles move from the header to SettingsFooter+Unsaved.swift
(guard-prover: both red-proofed); the warningInk wiring needle
repoints to SettingsHeaderBar+Status.swift, its surviving render
site.
