+++ Sources/KiwiDeskCore/Resources/Locales/en.json
@@ -322,6 +322,7 @@
   "footer.unsaved.count_one": "%1$d unsaved change",
   "footer.unsaved.count_one_to": "%1$d unsaved change to %2$@",
   "footer.unsaved.count_to": "%1$d unsaved changes to %2$@",
+  "footer.unsaved.help": "Click for the list of changes.",
   "footer.unsaved_changes": "Unsaved changes",
   "gaps.bottom": "Bottom",
   "gaps.horizontal": "Horizontal",
@@ -417,9 +418,6 @@
   "home.profiles.badge_ax": "start here",
   "home.title": "Settings",
   "home.title_ax": "KiwiDesk Settings",
-  "home.unsaved.count": "%1$d unsaved changes",
-  "home.unsaved.count_one": "1 unsaved change",
-  "home.unsaved.help": "Changes not saved yet — click for the list; Save and Revert are in the save pill.",
   "icon_picker.choose.help": "Choose an icon",
   "icon_picker.clear.help": "Remove icon (use the default)",
   "icon_picker.emoji": "Emoji",
