+++ Sources/KiwiDeskCore/Resources/Locales/fr.json
@@ -417,9 +417,6 @@
   "home.profiles.badge_ax": "commencez ici",
   "home.title": "Réglages",
   "home.title_ax": "Réglages KiwiDesk",
-  "home.unsaved.count": "%1$d modifications non enregistrées",
-  "home.unsaved.count_one": "1 modification non enregistrée",
-  "home.unsaved.help": "Modifications pas encore enregistrées — cliquez pour voir la liste ; Enregistrer et Annuler les modifications se trouvent en bas de la fenêtre.",
   "icon_picker.choose.help": "Choisir une icône",
   "icon_picker.clear.help": "Supprimer l'icône (utiliser celle par défaut)",
   "icon_picker.emoji": "Emoji",
