+++ Sources/KiwiDeskCore/Resources/Locales/it.json
@@ -417,9 +417,6 @@
   "home.profiles.badge_ax": "Inizia qui",
   "home.title": "Impostazioni",
   "home.title_ax": "Impostazioni di KiwiDesk",
-  "home.unsaved.count": "%1$d modifiche non salvate",
-  "home.unsaved.count_one": "1 modifica non salvata",
-  "home.unsaved.help": "Modifiche non ancora salvate: fai clic per l'elenco; Salva e Annulla modifiche si trovano nella barra di salvataggio in basso.",
   "icon_picker.choose.help": "Scegli un'icona",
   "icon_picker.clear.help": "Rimuovi l'icona (usa quella predefinita)",
   "icon_picker.emoji": "Emoji",
