+++ Sources/KiwiDeskCore/Resources/Locales/ja.json
@@ -417,9 +417,6 @@
   "home.profiles.badge_ax": "ここからスタート",
   "home.title": "設定",
   "home.title_ax": "KiwiDesk 設定",
-  "home.unsaved.count": "%1$d件の未保存の変更",
-  "home.unsaved.count_one": "1件の未保存の変更",
-  "home.unsaved.help": "変更はまだ保存されていません。クリックすると一覧が表示されます。「保存」と「変更を元に戻す」はウインドウ下部にあります。",
   "icon_picker.choose.help": "アイコンを選択",
   "icon_picker.clear.help": "アイコンを削除（デフォルトを使用）",
   "icon_picker.emoji": "絵文字",
