+++ Sources/KiwiDeskCore/Resources/Locales/ko.json
@@ -417,9 +417,6 @@
   "home.profiles.badge_ax": "여기서 시작",
   "home.title": "설정",
   "home.title_ax": "KiwiDesk 설정",
-  "home.unsaved.count": "%1$d개의 저장되지 않은 변경 사항",
-  "home.unsaved.count_one": "1개의 저장되지 않은 변경 사항",
-  "home.unsaved.help": "아직 저장하지 않은 변경 사항이 있습니다. 클릭하면 목록이 열립니다. ‘저장’과 ‘변경 사항 되돌리기’는 창 아래쪽에 떠 있는 막대에 있습니다.",
   "icon_picker.choose.help": "아이콘 선택",
   "icon_picker.clear.help": "아이콘 제거(기본값 사용)",
   "icon_picker.emoji": "이모티콘",
