+++ Sources/KiwiDeskCore/Resources/Locales/pt-BR.json
@@ -417,9 +417,6 @@
   "home.profiles.badge_ax": "Comece aqui",
   "home.title": "Ajustes",
   "home.title_ax": "Ajustes do KiwiDesk",
-  "home.unsaved.count": "%1$d alterações não salvas",
-  "home.unsaved.count_one": "1 alteração não salva",
-  "home.unsaved.help": "Alterações ainda não salvas — clique para ver a lista; “Salvar” e “Reverter alterações” ficam na barra flutuante de salvamento.",
   "icon_picker.choose.help": "Escolher um ícone",
   "icon_picker.clear.help": "Remover o ícone (usar o padrão)",
   "icon_picker.emoji": "Emoji",
