+++ Sources/KiwiDeskCore/Resources/Locales/ru.json
@@ -417,9 +417,6 @@
   "home.profiles.badge_ax": "начни здесь",
   "home.title": "Настройки",
   "home.title_ax": "Настройки KiwiDesk",
-  "home.unsaved.count": "Несохранённых изменений: %1$d",
-  "home.unsaved.count_one": "1 несохранённое изменение",
-  "home.unsaved.help": "Изменения ещё не сохранены — нажми, чтобы увидеть список; «Сохранить» и «Отменить изменения» — на плавающей панели внизу окна.",
   "icon_picker.choose.help": "Выбрать значок",
   "icon_picker.clear.help": "Убрать значок (использовать стандартный)",
   "icon_picker.emoji": "Эмодзи",
