+++ Sources/KiwiDeskCore/Resources/Locales/zh-Hans.json
@@ -417,9 +417,6 @@
   "home.profiles.badge_ax": "从这里开始",
   "home.title": "设置",
   "home.title_ax": "KiwiDesk 设置",
-  "home.unsaved.count": "%1$d 项未保存的更改",
-  "home.unsaved.count_one": "1 项未保存的更改",
-  "home.unsaved.help": "更改尚未保存 — 点按可查看列表；“保存”和“还原更改”位于浮动的保存条中。",
   "icon_picker.choose.help": "选择图标",
   "icon_picker.clear.help": "移除图标（使用默认图标）",
   "icon_picker.emoji": "表情符号",
