+++ Sources/KiwiDeskCore/Resources/Locales/zh-Hant.json
@@ -417,9 +417,6 @@
   "home.profiles.badge_ax": "從這裡開始",
   "home.title": "設定",
   "home.title_ax": "KiwiDesk 設定",
-  "home.unsaved.count": "%1$d 項未儲存的變更",
-  "home.unsaved.count_one": "1 項未儲存的變更",
-  "home.unsaved.help": "變更尚未儲存 — 點按可查看清單；「儲存」與「還原更動」在浮動的儲存列中。",
   "icon_picker.choose.help": "選擇圖像",
   "icon_picker.clear.help": "移除圖像（使用預設）",
   "icon_picker.emoji": "表情符號",
